+++ R/lamp-colors.r
@@ -1,22 +1,21 @@
 #' Spectral data for Lamps of different colours
 #' 
-#' Names of members of the collection of emission spectra
-#' \code{\link{lamps.mspct}} grouped by the wavelength at which they emit.
+#' Names of members of the \code{\link{lamps.mpsct}} collection of spectra
+#' grouped by colour or wavelength band.
 #' 
-#' @details
-#' The character vectors \code{"uv_leds"}, \code{"purple_leds"},
-#' \code{"blue_leds"}, \code{"green_leds"}, \code{"yellow_leds"},
-#' \code{"orange_leds"} and \code{"red_leds"} contain the names of the members
-#' of \code{lamps.mspct} with peaks of emission within the wavelength range
-#' corresponding to the light colours as defined by ISO standards. Vector
-#' \code{amber_lamps} is the union of \code{"yellow_lamps"} and
-#' \code{"orange_lamps"}. Vector \code{white_lamps} contains the names of spectra
-#' for lamps with broad or multiple peaks of emission in the visible range.
-#' Vectors \code{"uv_lamps"} and \code{"ir_lamps"} contain the names for lamps with
-#' peak emission at wavelengths < 400 nm and wavelengths > 700 nm, respectively.
-#' 
-#' These vectors can be used to extract subsets of spectra from
-#' \code{lamps.mspct}.
+#' @details contain the names of the members of \code{lamps.mspct} with peaks of
+#'   emission within the wavelength range corresponding to the light colours as
+#'   defined by ISO standards. Vector \code{amber_lamps} is the union of
+#'   \code{"yellow_lamps"} and \code{"orange_lamps"}. Vector \code{white_lamps}
+#'   contains the names of spectra for lamps with broad or multiple peaks of
+#'   emission in the visible range. Vectors \code{"uv_lamps"} and
+#'   \code{"ir_lamps"} contain the names for lamps with peak emission at
+#'   wavelengths < 400 nm and wavelengths > 700 nm, respectively.  One 
+#'   additional vector, \code{lamp_colors} contains the names of
+#'   the colors as used in the vectors.
+#'
+#'   These vectors can be used to extract subsets of spectra from
+#'   \code{lamps.mspct}.
 #' 
 #' @docType data
 #' @keywords datasets
@@ -24,7 +23,9 @@
 #' 
 #' @aliases 'lamps by color' 
 #' 
-#' @examples 
+#' @examples
+#' lamp_colors
+#' 
 #' uv_lamps
 #' blue_lamps
 #' red_lamps
@@ -44,9 +45,6 @@
 "purple_lamps"
 
 #' @rdname lamp_colors
-"ir_lamps"
-
-#' @rdname lamp_colors
 "blue_lamps"
 
 #' @rdname lamp_colors
@@ -60,6 +58,9 @@
 
 #' @rdname lamp_colors
 "red_lamps"
+
+#' @rdname lamp_colors
+"ir_lamps"
 
 #' @rdname lamp_colors
 "amber_lamps"