+++ R/lamp-types.r
@@ -1,23 +1,28 @@
 #' Spectral data for Lamps of different types
 #' 
-#' Names of datasets containing the wavelengths and tabulated values spectral emittance 
-#' for the lamps of different types. 
+#' Names of members of the \code{\link{lamps.mpsct}} collection of spectra
+#' grouped by the technology their are based on, i.e., type.
 #' 
-#' @note In the case of LED-based lamps we include here only ready built commercially 
-#' available lamps.
-#' In some cases assembled from multiple descrete LEDs, possible of mixed types and 
-#' spectral output. For emission spectra for LEDs available as electronic components
-#' please see \code{\link[photobiologyLEDs]{photobiologyLEDs-package}}.
+#' @details
+#'   These vectors can be used to extract subsets of spectra from
+#'   \code{lamps.mspct}.  One additional vector, \code{lamp_types} contains the 
+#'   names used for types of lamps in the names and vectors.
 #' 
-#' @aliases types
+#' @note In the case of LED-based lamps we include here only ready built
+#'   commercially available lamps. In some cases assembled from multiple
+#'   discrete LEDs, possible of mixed types and spectral output. For emission
+#'   spectra for LEDs available as electronic components please see
+#'   \code{\link[photobiologyLEDs]{photobiologyLEDs-package}}.
 #' 
 #' @docType data
 #' @keywords datasets
 #' @format A vector of character strings.
 #' 
-#' @aliases sodium_lamps mercury_lamps multimetal_lamps led_lamps
+#' @aliases 'lamps by technology'
 #' 
-#' @examples 
+#' @examples
+#' lamp_types
+#' 
 #' sodium_lamps
 #' multimetal_lamps
 #' mercury_lamps